+++ src/createMagicMoveComponent.js
@@ -6,18 +6,22 @@ import MagicMoveView from "./View";
  * Creates a magically moving component.
  *
  * @param {Component} Component - Source component
- * @param {Component} [AnimatedComponent] - The animated component
- * @param {string} [type] - Component type (e.g. "image", "text", etc..)
+ * @param {Object} [props] - Additional options
+ * @param {string} [props.ComponentType] - Component type (e.g. "image", "text", etc..)
+ * @param {Component} [props.AnimatedComponent] - The animated component
  */
-function createMagicMoveComponent(Component, AnimatedComponent, type) {
+function createMagicMoveComponent(Component, props) {
+  let { AnimatedComponent, ComponentType, ...otherProps } = props || {};
+  AnimatedComponent =
+    AnimatedComponent || Animated.createAnimatedComponent(Component);
+  ComponentType = ComponentType || "view";
   const magicMoveComponent = props => {
     return (
       <MagicMoveView
         Component={Component}
-        AnimatedComponent={
-          AnimatedComponent || Animated.createAnimatedComponent(Component)
-        }
-        ComponentType={type || "view"}
+        AnimatedComponent={AnimatedComponent}
+        ComponentType={ComponentType}
+        {...otherProps}
         {...props}
       />
     );